move all types/interfaces to src/types dir
