+++ src/fetchers/base.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { ZelConfig } from '../config';
+import type { ZelConfig } from '../types';
 import EventEmitter from 'events';
 import Promise from 'bluebird';
