+++ src/fetchers/file.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { ZelConfig } from '../config';
+import type { ZelConfig } from '../types';
 import BaseFetcher from './base';
 
 export default class FileFetcher extends BaseFetcher {
